Add server files
These are the files needed to to run the nodeJS server to query data from influxDB and show it in the browser.
Info: only the app.js and package.json files are important. node_modules can be installed with the command "npm install". The other .js files are debugs or not used in the main program